refactor: base component utils - move constant into helpers folders - move badge, button component into base folder - increase font size for ux improvements
